add back missing build-time TypeScript validation (#731)
* fix TypeScript errors in fp-ts code

Fixup on #718. It introduced new TypeScript errors,
but the corrected import statements expose type
information that was missing, exposing existing
errors, too.

* fix TypeScript error in maya form

Fix up on TCY changes in #694.

* add back missing build-time TypeScript validation

This used to be done in the old craco/webpack build system,
so add it to electron-vite, too.

* make TypeScript errors fatal to the build

noEmitOnError is surprisingly not redundant with noEmit,
despite the name.